cdk-backend/requirements.txt
aws-cdk-lib==2.146.0
constructs==10.4.2